security: authorize active profile actions without breaking API key auth
